refactor: split range getter into two parts
Split the function that returns the visual selection range into two parts that
return the start and the end of the range respectively, and use them in the
CLI-command composer. This is part of a plan to use a general object with
details about flags to compose the CLI command.